# core/docs/swagger.yaml
basePath: /
definitions:
...
  db.Monitor:
    properties:
      agent_id:
        type: string
      computed_health:
        description: Cached computed health (up | down | degraded | unknown)
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      health:
        description: up | down | degraded | unknown
        type: string
      id:
        type: string
      last_health_computation:
        description: When health was last computed
        type: string
      last_successful_report_at:
        type: string
      lifecycle:
        description: active | disabled | deleted
        type: string
      name:
        type: string
      reporting_interval_seconds:
        description: Monitor check interval
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
  db.MonitorReport:
    properties:
      collected_at:
        type: string
      created_at:
        type: string
      health:
        description: up | down
        type: string
      id:
        type: string
      monitor_id:
        type: string
      payload:
        type: string
    type: object
...
  service.RegisterMonitorRequest:
    properties:
      agent_id:
        type: string
      description:
        type: string
      last_checked:
        type: string
      name:
        type: string
      reporting_interval_seconds:
        description: Monitor check interval in seconds
        type: integer
      type:
        type: string
    required:
    - agent_id
    - description
    - last_checked
    - name
    - type
    type: object
  service.RegisterMonitorResponse:
    properties:
      monitor_id:
        type: string
    type: object
...
  service.UnregisterMonitorRequest:
    properties:
      agent_id:
        type: string
      monitor_id:
        type: string
    required:
    - agent_id
    - monitor_id
    type: object
  service.UnregisterMonitorResponse:
    properties:
      success:
        type: boolean
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8999
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Orion Core API for agent management and monitoring
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Orion Core API
  version: "1.0"
paths:
...
  /v1/agents/{agent_id}/register-monitor:
    post:
      consumes:
      - application/json
      description: Register a new monitor for a specific agent
      parameters:
      - description: Agent ID
        in: path
        name: agent_id
        required: true
        type: string
      - description: Monitor registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RegisterMonitorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.RegisterMonitorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Register a monitor
      tags:
      - monitors
...
  /v1/agents/{agent_id}/unregister-monitor:
    post:
      consumes:
      - application/json
      description: Unregister a monitor for a specific agent
      parameters:
      - description: Agent ID
        in: path
        name: agent_id
        required: true
        type: string
      - description: Monitor unregistration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UnregisterMonitorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.UnregisterMonitorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Unregister a monitor
      tags:
      - monitors
...
  /v1/agents/{id}/monitors:
    get:
      consumes:
      - application/json
      description: Get a paginated list of monitors for a specific agent with optional
        filters
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by health status (up|down|degraded|unknown)
        in: query
        name: health
        type: string
      - description: Filter by lifecycle status (active|disabled|deleted)
        in: query
        name: lifecycle
        type: string
      - default: 50
        description: Maximum number of monitors to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of monitors to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    count:
                      format: int64
                      type: integer
                    limit:
                      type: integer
                    monitors:
                      items:
                        $ref: '#/definitions/db.Monitor'
                      type: array
                    offset:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: List monitors for an agent
      tags:
      - monitors
...
  /v1/monitors/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific monitor including recent
        reports and computed health
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    computed_health:
                      type: string
                    monitor:
                      $ref: '#/definitions/db.Monitor'
                    recent_reports:
                      items:
                        $ref: '#/definitions/db.MonitorReport'
                      type: array
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get monitor details
      tags:
      - monitors
  /v1/monitors/{id}/history:
    get:
      consumes:
      - application/json
      description: Get a paginated list of reports for a specific monitor
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Maximum number of reports to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of reports to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    count:
                      format: int64
                      type: integer
                    limit:
                      type: integer
                    offset:
                      type: integer
                    reports:
                      items:
                        $ref: '#/definitions/db.MonitorReport'
                      type: array
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get monitor history
      tags:
      - monitors
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"